Misc. Read this description for theoretical not implemented notes
Pre-processing is done as part of building the component, which is done
in `buildComponent`, invoked right *after* `preBuildComponent runPreBuildHooks`
which executes the build rules.

Ignoring for now the problems from running the rules before pre-processing
and building the component...

The extra source build rules are determined at configure time (which is
ofc before pre-processing).

What's the problem?

Some of the extra sources are only determined after preprocessing
(function `preprocessExtras`), for example, a C file generated from
pre-processing a .hsc would be added to the component's C sources.

Because rules for extra sources are determined and even run before
preprocessing, how can we account for pre-process-generated-sources?

Another note is that we cannot check whether a file such as
Bar.hs, which is generated from Bar.hsc, exists at configure time,
because it will only be generated after pre-processing.

In any case, we may specify dependencies on files which, when compiling,
do not exist -- we'll fail accordingly.